Internalize more work
