+++ pkg/rsstorage/internal/integration_test/integration_test.go
@@ -37,7 +37,7 @@ func TestPackage(t *testing.T) { check.TestingT(t) }
 
 // This suite will be skipped when running tests with SQLite only. To test, use
 // the `test-integration` target. To run these tests only, use:
-// `MODULE=pkg/rsstorage/integration_test just test-integration -v github.com/rstudio/platform-lib/pkg/rsstorage/integration_test -check.f=StorageIntegrationSuite`
+// `MODULE=pkg/rsstorage/internal/integration_test just test-integration -v github.com/rstudio/platform-lib/pkg/rsstorage/internal/integration_test -check.f=StorageIntegrationSuite`
 type StorageIntegrationSuite struct {
 	pool          *pgxpool.Pool
 	tempdirhelper servertest.TempDirHelper
@@ -361,11 +361,11 @@ func (s *StorageIntegrationSuite) TestCopying(c *check.C) {
 // resolver failures.
 //
 // Run with (against local MinIO instance):
-//  `MODULE=pkg/rsstorage/integration_test just test-integration -v github.com/rstudio/platform-lib/pkg/rsstorage/integration_test -check.f=S3IntegrationSuite`
+//  `MODULE=pkg/rsstorage/internal/integration_test just test-integration -v github.com/rstudio/platform-lib/pkg/rsstorage/internal/integration_test -check.f=S3IntegrationSuite`
 //
 // To run against your own AWS S3 bucket:
 //   First, customize the variables in the test below as noted. Then, run:
-//     `MODULE=pkg/rsstorage/integration_test DEF_TEST_ARGS="-v" just test github.com/rstudio/platform-lib/pkg/rsstorage/integration_test -check.f=S3IntegrationSuite`
+//     `MODULE=pkg/rsstorage/internal/integration_test DEF_TEST_ARGS="-v" just test github.com/rstudio/platform-lib/pkg/rsstorage/internal/integration_test -check.f=S3IntegrationSuite`
 type S3IntegrationSuite struct{}
 
 var _ = check.Suite(&S3IntegrationSuite{})
